Combines txt2txt and img2img tabs
