+++ krita_stable_diffusion/interface/interfaces/panel.py
@@ -1,6 +1,5 @@
 from krita import *
 from krita_stable_diffusion.interface.tabs.configtab import ConfigTab
-from krita_stable_diffusion.interface.tabs.img2imgtab import Img2ImgTab
 from krita_stable_diffusion.interface.tabs.txt2imgtab import Txt2ImgTab
 from krita_stable_diffusion.interface.tabs.upscaletab import UpscaleTab
 
@@ -29,7 +28,7 @@ class KritaDockWidget(DockWidget):
     def create_interface(self):
         tabs = QTabWidget()
         tabs.addTab(*(Txt2ImgTab().tab()))
-        tabs.addTab(*(Img2ImgTab().tab()))
+        # tabs.addTab(*(Img2ImgTab().tab()))
         tabs.addTab(*(UpscaleTab().tab()))
         tabs.addTab(*(ConfigTab().tab()))
