+++ krita_stable_diffusion/interface/tabs/base.py
@@ -1,9 +1,8 @@
 import json
 import os
-
-from krita import *
 import logging
 import random
+from krita import *
 from krita_stable_diffusion.interface.interfaces.vertical_interface import VerticalInterface
 from krita_stable_diffusion.settings import APPLICATION_ID
